0.0.1-TEMP

+++ src/main/java/com/hml/admin/controller/RoleController.java
@@ -52,6 +52,10 @@ public class RoleController extends BaseController {
 	@PreAuthorize("hasAuthority('sys:role:delete')")
 	@PostMapping(value="/delete")
 	public HttpResult delete(@RequestBody Role record) {
+		Role role = roleService.getById(record.getId());
+		if(role.getName().equals(SysConstants.ADMIN)){
+			return HttpResult.error("超级管理员不能删除!");
+		}
 		return HttpResult.ok(roleService.removeById(record.getId()));
 	}
